[core] rename file synchronization.rs into updater.rs
